# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Swagger Movie-Api"
    },
    "host": "localhost:8081",
    "basePath": "/movie/api",
...
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "uikey": {
            "in": "header",
            "name": "uikey",
            "type": "apiKey"
        },
        "session_key": {
            "in": "header",
            "name": "apisessionkey",
            "type": "apiKey"
        }
    },
    "paths": {
...
        "/getInfolog": {
            "post": {
                "tags": [
                    "logs"
                ],
                "description": "retrieve info data logged in system",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "list of info data logged by date",
                        "schema": {
                            "$ref": "#/definitions/GetInfolog"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/GetInfolog"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [
                    {
                        "uikey": []
                    }
                ]
            }
        },
        "/getErrorlog": {
            "post": {
                "tags": [
                    "logs"
                ],
                "description": "retrieve error(s) logged in system",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "list of error(s) logged by date",
                        "schema": {
                            "$ref": "#/definitions/GetErrorlog"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/GetErrorlog"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [
                    {
                        "uikey": []
                    }
                ]
            }
        },
        "/getDebuglog": {
            "post": {
                "tags": [
                    "logs"
                ],
                "description": "retrieve debug data logged in system",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "list of debug data logged by date",
                        "schema": {
                            "$ref": "#/definitions/GetDebuglog"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/GetDebuglog"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [
                    {
                        "uikey": []
                    }
                ]
            }
        },
        "/getUser": {
            "post": {
                "tags": [
                    "user"
                ],
                "description": "user(s) profile in system",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "profile of user in system",
                        "schema": {
                            "$ref": "#/definitions/GetUser"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/GetUser"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [
                    {
                        "uikey": [],
                        "session_key": []
                    }
                ]
            }
        },
        "/createUser": {
            "post": {
                "tags": [
                    "user"
                ],
                "description": "user(s) profile in system",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "profile of user in system",
                        "schema": {
                            "$ref": "#/definitions/CreateUser"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/CreateUser"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [
                    {
                        "uikey": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "GetInfolog": {
            "type": "object",
            "required": [
                "startDate",
                "endDate"
            ],
            "properties": {
                "startDate": {
                    "type": "string",
                    "format": "date",
                    "uniqueItems": true
                },
                "endDate": {
                    "type": "string",
                    "format": "date",
                    "uniqueItems": true
                }
            }
        },
        "GetErrorlog": {
            "type": "object",
            "required": [
                "startDate",
                "endDate"
            ],
            "properties": {
                "startDate": {
                    "type": "string",
                    "format": "date",
                    "uniqueItems": true
                },
                "endDate": {
                    "type": "string",
                    "format": "date",
                    "uniqueItems": true
                }
            }
        },
        "GetDebuglog": {
            "type": "object",
            "required": [
                "startDate",
                "endDate"
            ],
            "properties": {
                "startDate": {
                    "type": "string",
                    "format": "date",
                    "uniqueItems": true
                },
                "endDate": {
                    "type": "string",
                    "format": "date",
                    "uniqueItems": true
                }
            }
        },
        "GetUser": {
            "type": "object",
            "required": [
                "userId"
            ],
            "properties": {
                "userId": {
                    "type": "string",
                    "uniqueItems": true
                }
            }
        },
        "CreateUser": {
            "type": "object",
            "required": [
                "userName",
                "password",
                "firstName",
                "lastName",
                "email",
                "address",
                "country"
            ],
            "properties": {
                "userName": {
                    "type": "string",
                    "uniqueItems": true
                },
                "password": {
                    "type": "string",
                    "uniqueItems": true
                },
                "firstName": {
                    "type": "string",
                    "uniqueItems": true
                },
                "lastName": {
                    "type": "string",
                    "uniqueItems": true
                },
                "email": {
                    "type": "string",
                    "uniqueItems": true
                },
                "address": {
                    "type": "string",
                    "uniqueItems": true
                },
                "country": {
                    "type": "string",
                    "uniqueItems": true
                }
            }
        },
...
    }
}